requirements.txt
requests==2.31.0
streamlit==1.28.0
plotly==5.17.0
python-dotenv==1.0.0
pytest==8.2.0
fastapi==0.115.6
uvicorn==0.32.1
crewai==0.41.1
langfuse
